find if SIZE_MAX is declared, and if it is use it for R_SIZE_T_MAX
git-svn-id: https://svn.r-project.org/R/trunk@30767 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/CommandLineArgs.c
@@ -199,8 +199,8 @@ R_common_command_line(int *pac, char **argv, Rstart Rp)
 				 "WARNING: %s value is invalid: ignored\n",
 				 *av);
 		    else
-			sprintf(msg, "WARNING: %s=%ld`%c': too large and ignored\n",
-				*av, value,
+			sprintf(msg, "WARNING: %s=%lu`%c': too large and ignored\n",
+				*av, (unsigned long) value,
 				(ierr == 1) ? 'M': ((ierr == 2) ? 'K' : 'k'));
 		    R_ShowMessage(msg);
